astakos: Fix sqlite-related migration issues
As per 63bcab0, do not attempt to delete index when run on sqlite.

+++ snf-astakos-app/astakos/im/migrations/0044_auto__chg_field_projectapplication_chain__chg_field_project_id.py
@@ -14,8 +14,9 @@ class Migration(SchemaMigration):
         # Changing field 'Project.id'
         db.alter_column('im_project', 'id', self.gf('django.db.models.fields.BigIntegerField')(primary_key=True, db_column='id'))
 
-        # Removing index on 'Project', fields ['id']
-        db.delete_index('im_project', ['id'])
+        if db.backend_name != 'sqlite3':
+            # Removing index on 'Project', fields ['id']
+            db.delete_index('im_project', ['id'])
     
     
     def backwards(self, orm):
